Refactor ProspectSource resource structure: migrate to new namespace and implement resource pages
